// libs/enums/product.enum.ts
export enum ProductType {
  FOOD = "FOOD",
  TOY = "TOY",
  CLOTHES = "CLOTHES",
  HEALTH = "HEALTH",
  ACCESSORIES = "ACCESSORIES",
  OTHER = "OTHER",
}

export enum ProductStatus {
  ACTIVE = "ACTIVE",
  PAUSE = "PAUSE",
  DELETE = "DELETE",
}

export enum ProductPetType {
  DOG = "DOG",
  CAT = "CAT",
  RABBIT = "RABBIT",
  BIRD = "BIRD",
  HAMSTER = "HAMSTER",
  OTHER = "OTHER",
}